Cohorts section updated in index.tsx
Updated the Cohorts section replacing the table by a new image. This is my first code change PR (& also writing in a .tsx file) so I hope I didn't mess up the divs annotation or broke anything!

I think there's an extra div that we don't need now, that was probably centering the text column with the table or smth, and now that we don't have the table anymore it's probably redundant, but not sure.